LangGraph_projects/Chatbot_advanced/pyproject.toml
[project]
name = "chatbot-advanced"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.9.0",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
